services/ai-interview/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
livekit-server-sdk-python==1.3.0
pydantic==2.6.4
httpx==0.27.0
python-multipart==0.0.9
websockets==12.0